Moved handler final logic to a separate file.
