+++ components/funder/src/handler/handle_control.rs
@@ -12,7 +12,7 @@ use proto::funder::messages::{FriendStatus, UserRequestSendFunds,
     FunderOutgoingControl, ResponseSendFundsResult};
 
 use crate::ephemeral::Ephemeral;
-use crate::handler::{MutableFunderState, MutableEphemeral, is_friend_ready};
+use crate::handler::handler::{MutableFunderState, MutableEphemeral, is_friend_ready};
 use crate::handler::sender::SendCommands;
 use crate::handler::canceler::{cancel_local_pending_requests, 
     cancel_pending_user_requests, cancel_pending_requests};
