+++ components/funder/src/handler/handle_init.rs
@@ -3,7 +3,7 @@ use proto::funder::messages::FriendStatus;
 
 use crate::types::{ChannelerConfig, ChannelerAddFriend};
 
-use super::MutableFunderState;
+use crate::handler::handler::MutableFunderState;
 
 pub fn handle_init<A>(m_state: &MutableFunderState<A>,
                       outgoing_channeler_config: &mut Vec<ChannelerConfig<A>>)
